requirements.txt
streamlit==1.31.0
openai==1.12.0
python-dotenv==1.0.0
requests==2.31.0
pyyaml==6.0.1
python-dateutil==2.8.2
pytz==2024.1
emoji==2.8.0 